+++ include/pal/pal_core.h
@@ -411,71 +411,69 @@ static inline PalResult PAL_CALL palMakeResult(
 }
 
 /**
- * @brief Combine two 32-bit unsigned integers into a single 64-bit signed
- * integer.
+ * @brief Combine two 32-bit unsigned integers into a single 64-bit unsigned integer.
  *
- * @return The combined 64-bit signed integer.
+ * @return The combined 64-bit unsigned integer.
  *
  * Thread safety: Thread safe.
  *
  * @since 1.0
  * @sa palUnpackUint32
  */
-static inline int64_t PAL_CALL palPackUint32(
+static inline uint64_t PAL_CALL palPackUint32(
     uint32_t low,
     uint32_t high)
 {
-    return (int64_t)(((uint64_t)high << 32) | (uint64_t)low);
+    return (uint64_t)(((uint64_t)high << 32) | (uint64_t)low);
 }
 
 /**
- * @brief Combine two 32-bit signed integers into a single 64-bit signed
- * integer.
+ * @brief Combine two 32-bit signed integers into a single 64-bit unsigned integer.
  *
- * @return The combined 64-bit signed integer.
+ * @return The combined 64-bit unsigned integer.
  *
  * Thread safety: Thread safe.
  *
  * @since 1.0
  * @sa palUnpackInt32
  */
-static inline int64_t PAL_CALL palPackInt32(
+static inline uint64_t PAL_CALL palPackInt32(
     int32_t low,
     int32_t high)
 {
-    return ((int64_t)(uint32_t)high << 32) | (uint32_t)low;
+    return ((uint64_t)(uint32_t)high << 32) | (uint32_t)low;
 }
 
 /**
- * @brief Pack a pointer into a 64-bit signed integer.
+ * @brief Pack a pointer into a 64-bit unsigned integer.
  *
- * @return The packed 64-bit signed integer.
+ * @return The packed 64-bit unsigned integer.
  *
  * Thread safety: Thread safe.
  *
  * @since 1.0
  * @sa palUnpackPointer
  */
-static inline int64_t PAL_CALL palPackPointer(void* ptr)
+static inline uint64_t PAL_CALL palPackPointer(void* ptr)
 {
-    return (int64_t)(uintptr_t)ptr;
+    return (uint64_t)(uintptr_t)ptr;
 }
 
 /**
- * @brief Combine two floats into a single 64-bit signed integer.
+ * @brief Combine two floats into a single 64-bit unsigned integer.
  *
- * @return The combined 64-bit signed integer.
+ * @return The combined 64-bit unsigned integer.
  *
  * Thread safety: Thread safe.
  *
  * @since 1.3
  * @sa palUnpackFloat
  */
-static inline int64_t PAL_CALL palPackFloat(
+static inline uint64_t PAL_CALL palPackFloat(
     float low,
     float high)
 {
-    int64_t combined = 0;
+    uint64_t combined = 0;
 #if PAL_BIG_ENDIAN
     memcpy(&((uint32_t*)&combined)[0], &high, sizeof(float));
     memcpy(&((uint32_t*)&combined)[1], &low, sizeof(float));
@@ -488,10 +486,10 @@ static inline int64_t PAL_CALL palPackFloat(
 }
 
 /**
- * @brief Retrieve two 32-bit unsigned integers from a 64-bit signed integer.
+ * @brief Retrieve two 32-bit unsigned integers from a 64-bit unsigned integer.
  *
- * @param[out] outLow Low value of the 64-bit signed integer.
- * @param[out] outHigh High value of the 64-bit signed integer.
+ * @param[out] outLow Low value of the 64-bit unsigned integer.
+ * @param[out] outHigh High value of the 64-bit unsigned integer.
  *
  * Thread safety: Thread safe.
  *
@@ -499,7 +497,7 @@ static inline int64_t PAL_CALL palPackFloat(
  * @sa palPackUint32
  */
 static inline void PAL_CALL palUnpackUint32(
-    int64_t data,
+    uint64_t data,
     uint32_t* outLow,
     uint32_t* outHigh)
 {
@@ -513,10 +511,10 @@ static inline void PAL_CALL palUnpackUint32(
 }
 
 /**
- * @brief Retrieve two 32-bit signed integers from a 64-bit signed integer.
+ * @brief Retrieve two 32-bit signed integers from a 64-bit unsigned integer.
  *
- * @param[out] outLow Low value of the 64-bit signed integer.
- * @param[out] outHigh High value of the 64-bit signed integer.
+ * @param[out] outLow Low value of the 64-bit unsigned integer.
+ * @param[out] outHigh High value of the 64-bit unsigned integer.
  *
  * Thread safety: Thread-safe if `outLow` and `outHigh` are
  * thread local.
@@ -525,7 +523,7 @@ static inline void PAL_CALL palUnpackUint32(
  * @sa palPackInt32
  */
 static inline void PAL_CALL palUnpackInt32(
-    int64_t data,
+    uint64_t data,
     int32_t* outLow,
     int32_t* outHigh)
 {
@@ -539,25 +537,25 @@ static inline void PAL_CALL palUnpackInt32(
 }
 
 /**
- * @brief Unpack a pointer from a 64-bit signed integer.
+ * @brief Unpack a pointer from a 64-bit unsigned integer.
  *
- * @return The pointer from the 64-bit signed integer.
+ * @return The pointer from the 64-bit unsigned integer.
  *
  * Thread safety: Thread safe.
  *
  * @since 1.0
  * @sa palPackPointer
  */
-static inline void* PAL_CALL palUnpackPointer(int64_t data)
+static inline void* PAL_CALL palUnpackPointer(uint64_t data)
 {
     return (void*)(uintptr_t)data;
 }
 
 /**
- * @brief Retrieve two floats from a 64-bit signed integer.
+ * @brief Retrieve two floats from a 64-bit unsigned integer.
  *
- * @param[out] outLow Low value of the 64-bit signed integer.
- * @param[out] outHigh High value of the 64-bit signed integer.
+ * @param[out] outLow Low value of the 64-bit unsigned integer.
+ * @param[out] outHigh High value of the 64-bit unsigned integer.
  *
  * Thread safety: Thread-safe if `outLow` and `outHigh` are
  * thread local.
@@ -566,7 +564,7 @@ static inline void* PAL_CALL palUnpackPointer(int64_t data)
  * @sa palPackFloat
  */
 static inline void PAL_CALL palUnpackFloat(
-    int64_t data,
+    uint64_t data,
     float* low,
     float* high)
 {